recursive folder delete (ustable, retest)

+++ BigTreeFS.js
@@ -1,6 +1,5 @@
 const fs = require("fs");
 const stream = require("stream");
-const util = require("util");
 const auth = require("./AccessControlService");
 const combStream = require("combined-stream");
 const escTransform = require("./EscapeTransformStream");
@@ -21,19 +20,37 @@ this.readPath = function(path, jwt) {
 
 this.writePath = function(path, jwt) {
     if(!auth.isAuthorisedOperation(path, this.WRITE_OPERATION, jwt)) throw "access denied";
+
+    if(fs.existsSync(filePath(path))) fs.unlinkSync(filePath(path));
+
     var writable = fs.createWriteStream(filePath(path));
     return writable;
 };
 
 this.deletePath = function(path, jwt) {
     if(!auth.isAuthorisedOperation(path, this.DELETE_OPERATION, jwt)) throw "access denied";
-    fs.unlinkSync(filePath(path));
+
+    var fileName = filePath(path);
+    console.log("deleting ", fileName);
+    var isDirectory = fs.statSync(fileName).isDirectory();
+    console.log(isDirectory?"directory delete":"file delete");
+    if(isDirectory) {
+        rmdir(fileName);
+    } else {
+        fs.unlinkSync(fileName);
+    }
 };
 
 function filePath(path) {
     return process.env[(process.platform == 'win32') ? 'USERPROFILE' : 'HOME']+"/.bigTree"+path;
 }
 
+/**
+ * Returns a readable stream that contain a folder and its sub folders as a JSON object with files and their contents
+ * as fields and values
+ *
+ * @param path - absolute path on the local drive
+ */
 function dirAsJSON(path) {
     var comb = combStream.create();
     comb.append("{");
@@ -50,11 +67,28 @@ function dirAsJSON(path) {
     }
     comb.append("}");
     return comb;
-}
+};
 
 function fileAsJSON(path) {
     return fs.createReadStream(path).pipe(escTransform());
-}
+};
 
+function rmdir(path) {
+    if( fs.existsSync(path) ) {
+        console.log("file exists, deleting ", path);
+        fs.readdirSync(path).forEach(function(file,index) {
+            var curPath = path + "/" + file;
+            if(fs.lstatSync(curPath).isDirectory()) { // recurse
+                rmdir(curPath);
+            } else { // delete file
+                fs.unlinkSync(curPath);
+            }
+        });
+        console.log("rmdirSync ", path);
+        fs.rmdirSync(path);
+    }
+};
 
 //dirAsJSON("bigTree").pipe(process.stdout)
+
+//console.log(fs.rmdirSync(filePath("/bigTree/aaa/bbb/aaa/drei/ein")));
